% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CarEmission.R
\docType{package}
\name{CarEmission}
\alias{CarEmission}
\title{CarEmission: Pakiet selekcji i obliczania emisji spalin dla wybranych danych oraz ich wizualizacji}
\description{
Pakiet składa się z 2 funkcji, do selekcji i prezentacji graficznej danych dotyczcych emisji zanieczyszczeń dostepnych na stronie
Europejskiej Agencji Środowiska EEA
\url{https://www.eea.europa.eu/}
}
