% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Emission.R
\name{Emission}
\alias{Emission}
\title{Obliczenia emisji spalin}
\usage{
Emission(
  dane = input,
  kategoria = "Passenger Cars",
  euro = "Euro 5",
  mode = "",
  substancja = c("CO", "EC")
)
}
\arguments{
\item{dane}{dataframe}

\item{kategoria}{character}

\item{euro}{character}

\item{mode}{character}

\item{substancja}{character}
}
\value{
dataframe
}
\description{
Obliczenia emisji spalin
}
\examples{
\donttest{Emission()}
}
