% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Emsplot.R
\name{Emsplot}
\alias{Emsplot}
\title{Stworzenie wykresu}
\usage{
Emsplot(type, out, x, y, z, title)
}
\arguments{
\item{out}{double}

\item{x}{double}

\item{y}{double}
}
\value{
plot
}
\description{
Stworzenie wykresu
}
