% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\docType{data}
\name{input}
\alias{input}
\title{Dane wejsciowe używane dla funkcji szacowania emisji:}
\format{
Ramka danych zawiera:
\describe{
\item{Nat}{Natezenie ruchu wygenerowane losowo}
\item{Segment}{Segment pojazdow - Mini itd.}
\item{Fuel}{Rodzaj paliwa - Petrol itd.}
\item{Technology}{Technologia pracy silnika - GDI itd.}
}
}
\usage{
input
}
\description{
Dane wejsciowe używane dla funkcji szacowania emisji:
}
\examples{
\dontrun{
 input
 
}
}
\keyword{datasets}
