% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Emission.R
\name{Emission}
\alias{Emission}
\title{Funkcja emisje spalin na podstawie danych zawartych w ramkach danych "wskazniki" i "input"}
\usage{
Emission(
  dane = input,
  kategoria = "Passenger Cars",
  euro = c("Euro 4", "Euro 5"),
  mode = "",
  substancja = c("CO", "EC")
)
}
\arguments{
\item{dane}{dataframe}

\item{kategoria}{character}

\item{euro}{character}

\item{mode}{character}

\item{substancja}{character}
}
\value{
dataframe
}
\description{
Algorytm obliczen emisji:
Emisja = Nat x ((Alpha  x Procent^2 + Beta x Procent + Gamma + (Delta/Procent)) / (Epsilon x Procent^2 + Zita x Procent + Hta) x (1- Reduction))
}
\examples{
\donttest{Emission()}
}
