% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Emsplot.R
\name{Emsplot}
\alias{Emsplot}
\title{Funkcja do wizualizacji danych na podstawie danych wygenerowanych przy pomocy funkcji Emission}
\usage{
Emsplot(type, data, x, y, z, title)
}
\arguments{
\item{type}{text - Typ wykresu}

\item{data}{double - dane}

\item{x}{double - dane na osi x}

\item{y}{double - dane na osi y}

\item{z}{double}

\item{title}{text - Tytul}
}
\value{
plot
}
\description{
Funkcja do wizualizacji danych na podstawie danych wygenerowanych przy pomocy funkcji Emission
}
\examples{
\donttest{Emsplot("Example")}
}
