% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Emsplot.R
\name{Emsplot}
\alias{Emsplot}
\title{Stworzenie wykresu}
\usage{
Emsplot(type, data, x, y, z, title)
}
\arguments{
\item{type}{text}

\item{data}{double}

\item{x}{double}

\item{y}{double}

\item{z}{doubl}

\item{title}{text}
}
\value{
plot
}
\description{
Stworzenie wykresu
}
\examples{
examples \donttest{Emsplot("Example")}
}
